refactor: deduplicate sources_by_id into serde_helpers
Move the sources_by_id serialization module to a single location in
foundry_compilers_artifacts::serde_helpers and use it from both:
- foundry_compilers_artifacts::CompilerOutput
- foundry_compilers::compilers::CompilerOutput

This removes ~110 lines of duplicated code.
