+++ crates/artifacts/solc/src/lib.rs
@@ -16,61 +16,6 @@ use std::{
     str::FromStr,
 };
 
-mod sources_by_id {
-    use crate::SourceFile;
-    use serde::{
-        de::{MapAccess, Visitor},
-        ser::SerializeMap,
-        Deserializer, Serializer,
-    };
-    use std::{collections::BTreeMap, fmt, path::PathBuf};
-
-    pub fn serialize<S>(
-        sources: &BTreeMap<PathBuf, SourceFile>,
-        serializer: S,
-    ) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let mut entries: Vec<_> = sources.iter().collect();
-        entries.sort_by_key(|(_, source)| source.id);
-
-        let mut map = serializer.serialize_map(Some(entries.len()))?;
-        for (path, source) in entries {
-            map.serialize_entry(path, source)?;
-        }
-        map.end()
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<BTreeMap<PathBuf, SourceFile>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct SourcesVisitor;
-
-        impl<'de> Visitor<'de> for SourcesVisitor {
-            type Value = BTreeMap<PathBuf, SourceFile>;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-                formatter.write_str("a map of source files")
-            }
-
-            fn visit_map<M>(self, mut access: M) -> Result<Self::Value, M::Error>
-            where
-                M: MapAccess<'de>,
-            {
-                let mut map = BTreeMap::new();
-                while let Some((key, value)) = access.next_entry::<PathBuf, SourceFile>()? {
-                    map.insert(key, value);
-                }
-                Ok(map)
-            }
-        }
-
-        deserializer.deserialize_map(SourcesVisitor)
-    }
-}
-
 pub mod error;
 pub use error::*;
 pub mod ast;
@@ -1529,7 +1474,7 @@ pub struct DocLibraries {
 pub struct CompilerOutput {
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub errors: Vec<Error>,
-    #[serde(default, with = "sources_by_id")]
+    #[serde(default, with = "serde_helpers::sources_by_id")]
     pub sources: BTreeMap<PathBuf, SourceFile>,
     #[serde(default)]
     pub contracts: Contracts,
